# apps/service/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Me": {
        "properties": {
          "active_scenario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Scenario"
          },
          "book_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Id"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "has_book": {
            "title": "Has Book",
            "type": "boolean"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "email",
          "created_at",
          "has_book"
        ],
        "title": "Me",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Deterministic service core for the CashKit consumer MLP (SPEC-mlp-consumer.md). Every money figure is an engine number, serialized once, and every payload that carries one carries its provenance with it.",
    "title": "CashKit MLP service",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/me": {
      "delete": {
        "description": "Full account deletion (SPEC \u00a79).\n\nFour things go, and the fourth is the one a cascade cannot reach.\n\n1. **Every session**, so no device keeps reading.\n2. **The book directory**, off the volume, immediately.\n3. **Every Postgres row for the account** \u2014 ``turns``, ``llm_calls``,\n   ``proposals`` and ``import_jobs`` included, by cascade from ``users``,\n   which is the root of all of them. The row itself goes too, rather than\n   gaining a ``deleted_at``: keeping it would keep the email for ever and\n   would block the address from ever signing up again (D-MLP-22).\n4. **Every magic-link token issued to the address.** ``login_tokens`` has\n   no ``user_id`` \u2014 a link can be requested before an account exists \u2014 so\n   no cascade reaches it, and an unconsumed row would leave the address in\n   the database after the account that owned it was erased. It is deleted\n   by address, which also burns any link already in flight.\n\nBackups cannot be immediate: an object already in the bucket holds the\naccount until it ages out. SPEC \u00a79 gives that thirty days, and the\n``deletions`` receipt is what carries the obligation once the row that\nowed it is gone. It holds no personal data \u2014 a uuid that now references\nnothing, and two timestamps. :mod:`cashkit_service.retention` closes it\nagainst what the bucket actually still holds, not against elapsed time.",
        "operationId": "delete_me_me_delete",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Me",
        "tags": [
          "account"
        ]
      },
      "get": {
        "operationId": "get_me_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Me",
        "tags": [
          "account"
        ]
      }
    },
    "/me/export": {
      "get": {
        "description": "Everything the user owns, one archive (GDPR, SPEC \u00a73/\u00a79).\n\nPostgres rows plus the book directory itself \u2014 the YAML revisions and the\nledger are the user's data, not an internal format they are locked out of.\nSession and link token hashes are excluded: they are credentials, and\nexporting them would widen the blast radius of a leaked archive.",
        "operationId": "export_me_me_export_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/zip": {}
            },
            "description": "Everything the account owns."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Me",
        "tags": [
          "account"
        ]
      }
    },
...
  }
}